connector/test_talk/connector/requirements.txt
tinvest
kafka-python
google
protobuf